Adding a disassembler

+++ emu/src/breakpoints.rs
@@ -11,7 +11,7 @@ pub enum BreakPointTypes {
 
 #[derive(Clone, Debug, PartialEq, PartialOrd, Copy)]
 pub struct BreakPoint {
-    addr: u16,
+    addr: usize,
     bp_type: BreakPointTypes,
     active: bool,
     id: usize,
@@ -19,7 +19,7 @@ pub struct BreakPoint {
 
 impl BreakPoint {
     /// Describes a breakpoint constructs defaulted to active
-    pub fn new(bp_type: BreakPointTypes, addr: u16, id: usize) -> BreakPoint {
+    pub fn new(bp_type: BreakPointTypes, addr: usize, id: usize) -> BreakPoint {
         BreakPoint {
             bp_type,
             addr,
@@ -67,12 +67,12 @@ impl BreakPoints {
         self.break_points.len()
     }
 
-    pub fn has_any_breakpoint(&self, addr: u16) -> bool {
+    pub fn has_any_breakpoint(&self, addr: usize) -> bool {
         let bp = self.get_breakpoints(addr, 1);
         bp.iter().filter(|b| b.addr == addr).count() > 0
     }
 
-    pub fn has_breakpoint(&self, addr: u16, bp_type: BreakPointTypes) -> bool {
+    pub fn has_breakpoint(&self, addr: usize, bp_type: BreakPointTypes) -> bool {
         let bp = self.get_breakpoints(addr, 1);
         let num = bp
             .iter()
@@ -81,7 +81,7 @@ impl BreakPoints {
         num > 0
     }
 
-    pub fn add(&mut self, addr: u16, bp_type: BreakPointTypes) -> Option<usize> {
+    pub fn add(&mut self, addr: usize, bp_type: BreakPointTypes) -> Option<usize> {
         if !self.has_breakpoint(addr, bp_type) {
             let ret = self.id;
             let bp = BreakPoint::new(bp_type, addr, ret);
@@ -92,11 +92,11 @@ impl BreakPoints {
             None
         }
     }
-    pub fn find_breakpoint_id(&self, addr: u16, bp_type: BreakPointTypes) -> Option<usize> {
+    pub fn find_breakpoint_id(&self, addr: usize, bp_type: BreakPointTypes) -> Option<usize> {
         self.find_breakpoint(addr, bp_type).map(|bp| bp.id)
     }
 
-    fn find_breakpoint(&self, addr: u16, bp_type: BreakPointTypes) -> Option<&BreakPoint> {
+    fn find_breakpoint(&self, addr: usize, bp_type: BreakPointTypes) -> Option<&BreakPoint> {
         for (_, bp) in self.break_points.iter() {
             if bp.addr == addr && bp.bp_type == bp_type {
                 return Some(bp);
@@ -108,17 +108,17 @@ impl BreakPoints {
         self.break_points.remove(&id);
     }
 
-    pub fn remove(&mut self, addr: u16, bp_type: BreakPointTypes) {
+    pub fn remove(&mut self, addr: usize, bp_type: BreakPointTypes) {
         if let Some(id) = self.find_breakpoint_id(addr, bp_type) {
             self.break_points.remove(&id);
         }
     }
 
-    fn get_range(addr: u16, range: usize) -> Result<Region, RegionErr> {
+    fn get_range(addr: usize, range: usize) -> Result<Region, RegionErr> {
         Region::checked_new(addr, range)
     }
 
-    pub fn get_breakpoints(&self, addr: u16, range: usize) -> Vec<&BreakPoint> {
+    pub fn get_breakpoints(&self, addr: usize, range: usize) -> Vec<&BreakPoint> {
         if let Ok(r) = Self::get_range(addr, range) {
             self.break_points
                 .values()
@@ -129,7 +129,7 @@ impl BreakPoints {
         }
     }
 
-    pub fn get_breakpoints_mut(&mut self, addr: u16, range: usize) -> Vec<&mut BreakPoint> {
+    pub fn get_breakpoints_mut(&mut self, addr: usize, range: usize) -> Vec<&mut BreakPoint> {
         if let Ok(r) = Self::get_range(addr, range) {
             self.break_points
                 .values_mut()
@@ -140,7 +140,7 @@ impl BreakPoints {
         }
     }
 
-    pub fn remove_all_at_addr(&mut self, addr: u16) {
+    pub fn remove_all_at_addr(&mut self, addr: usize) {
         let v: Vec<usize> = self.get_breakpoints(addr, 1).iter().map(|b| b.id).collect();
 
         for id in v {
