+++ emu/src/mem/lmemmap.rs
@@ -5,14 +5,14 @@ use std::fmt;
 
 #[derive(Debug, Clone, Default)]
 pub struct LogEntry {
-    pub addr: u16,
+    pub addr: usize,
     pub write: bool,
     pub val: u16,
     pub word: bool,
 }
 
 impl LogEntry {
-    fn write_byte(addr: u16, val: u8) -> LogEntry {
+    fn write_byte(addr: usize, val: u8) -> LogEntry {
         LogEntry {
             addr,
             write: true,
@@ -21,7 +21,7 @@ impl LogEntry {
         }
     }
 
-    fn read_byte(addr: u16, val: u8) -> LogEntry {
+    fn read_byte(addr: usize, val: u8) -> LogEntry {
         LogEntry {
             addr,
             write: false,
@@ -30,7 +30,7 @@ impl LogEntry {
         }
     }
 
-    fn write_word(addr: u16, val: u16) -> LogEntry {
+    fn write_word(addr: usize, val: u16) -> LogEntry {
         LogEntry {
             addr,
             write: true,
@@ -39,7 +39,7 @@ impl LogEntry {
         }
     }
 
-    fn read_word(addr: u16, val: u16) -> LogEntry {
+    fn read_word(addr: usize, val: u16) -> LogEntry {
         LogEntry {
             addr,
             write: false,
@@ -110,7 +110,7 @@ impl MemoryIO for LoggingMemMap {
         self.mem_map.update_sha1(digest)
     }
 
-    fn upload(&mut self, addr: u16, data: &[u8]) -> MemResult<()> {
+    fn upload(&mut self, addr: usize, data: &[u8]) -> MemResult<()> {
         self.mem_map.upload(addr, data)
     }
 
@@ -122,28 +122,28 @@ impl MemoryIO for LoggingMemMap {
         self.mem_map.get_range()
     }
 
-    fn load_byte(&mut self, addr: u16) -> MemResult<u8> {
+    fn load_byte(&mut self, addr: usize) -> MemResult<u8> {
         let val = self.mem_map.load_byte(addr)?;
         let msg = LogEntry::read_byte(addr, val);
         self.log(msg);
         Ok(val)
     }
 
-    fn store_byte(&mut self, addr: u16, val: u8) -> MemResult<()> {
+    fn store_byte(&mut self, addr: usize, val: u8) -> MemResult<()> {
         self.mem_map.store_byte(addr, val)?;
         let msg = LogEntry::write_byte(addr, val);
         self.log(msg);
         Ok(())
     }
 
-    fn store_word(&mut self, addr: u16, val: u16) -> MemResult<()> {
+    fn store_word(&mut self, addr: usize, val: u16) -> MemResult<()> {
         self.mem_map.store_word(addr, val)?;
         let msg = LogEntry::write_word(addr, val);
         self.log(msg);
         Ok(())
     }
 
-    fn load_word(&mut self, addr: u16) -> MemResult<u16> {
+    fn load_word(&mut self, addr: usize) -> MemResult<u16> {
         let val = self.mem_map.load_word(addr)?;
         let msg = LogEntry::read_word(addr, val);
         self.log(msg);
